src/rag_open_source/requirements.txt
absl-py==2.1.0
accelerate==0.28.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asgiref==3.8.1
astrapy==0.7.7
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bitsandbytes==0.43.0
black==24.3.0
bs4==0.0.2
build==1.2.1
cachetools==5.3.3
cassandra-driver==3.29.1
cassio==0.1.5
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chevron==0.14.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cloudpickle==3.0.0
cmake==3.29.0.1
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
dataclasses-json==0.6.4
datasets==2.18.0
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@ff53992b1985b63bd3262b5a36167098e3dada02
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
evaluate==0.4.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
filelock==3.13.3
FlagEmbedding @ git+https://github.com/FlagOpen/FlagEmbedding.git@46062d983a334614ea5d41168274e7acb152c06e
flatbuffers==24.3.25
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.2.0
fvcore==0.1.5.post20221221
geomet==0.2.1.post1
google-api-core==1.34.1
google-auth==2.29.0
google-cloud-vision==1.0.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.48.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.1
humanfriendly==10.0
hydra-core==1.3.2
hyperframe==6.0.1
idna==3.6
img2pdf==0.5.1
img2table==1.2.11
importlib-metadata==6.11.0
importlib_resources==6.4.0
iopath==0.1.9
ipykernel==6.29.4
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
JPype1==1.5.0
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
kubernetes==29.0.0
langfuse==2.21.1
layoutparser==0.3.4
llama-index==0.10.20
llama-index-agent-openai==0.1.7
llama-index-callbacks-langfuse==0.1.3
llama-index-cli==0.1.11
llama-index-core==0.10.20.post1
llama-index-embeddings-huggingface==0.2.0
llama-index-embeddings-openai==0.1.7
llama-index-extractors-entity==0.1.2
llama-index-indices-managed-llama-cloud==0.1.5
llama-index-legacy==0.9.48
llama-index-llms-llama-cpp==0.1.3
llama-index-llms-openai==0.1.13
llama-index-multi-modal-llms-openai==0.1.4
llama-index-postprocessor-flag-embedding-reranker==0.1.2
llama-index-program-openai==0.1.5
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.12
llama-index-readers-llama-parse==0.1.4
llama-index-vector-stores-astra-db==0.1.5
llama-index-vector-stores-chroma==0.1.6
llama-index-vector-stores-mongodb==0.1.4
llama-index-vector-stores-pinecone==0.1.4
llama-parse==0.4.0
llama_cpp_python==0.2.57
llamaindex-py-client==0.1.15
llvmlite==0.42.0
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
minijinja==1.0.16
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
ocrmypdf==16.1.2
omegaconf==2.3.0
onnxruntime==1.17.1
openai==1.14.3
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.10.0
overrides==7.7.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pexpect==4.9.0
pikepdf==8.14.0
pillow==10.2.0
pinecone-client==3.2.1
pkgconfig==1.5.5
platformdirs==4.2.0
pluggy==1.4.0
polars==0.20.16
poppler-utils==0.1.0
portalocker==2.8.2
posthog==3.5.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pymongo==4.6.3
PyMuPDF==1.24.0
PyMuPDFb==1.24.0
pyparsing==3.1.2
pypdf==4.1.0
pypdfium2==4.28.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
rag-open-source @ file:///home/giacomo/Documenti/GitHub/RAG-with-LLamaIndex-Open/dist/rag_open_source-0.1-py3-none-any.whl
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
responses==0.18.0
rich==13.7.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
sentence-transformers==2.6.1
seqeval==1.2.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
span-marker==1.5.0
SQLAlchemy==2.0.29
src @ file:///home/giacomo/Documenti/GitHub/RAG-with-LLamaIndex-Open/dist/src-0.1-py3-none-any.whl
stack-data==0.6.3
starlette==0.36.3
striprtf==0.0.26
sympy==1.12
tabula-py==2.9.0
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.4.0
tiktoken==0.6.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.39.1
triton==2.2.0
typer==0.11.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
XlsxWriter==3.2.0
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.4
zipp==3.18.1
